requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
email-validator
stellar-sdk